CN-8 feat: s3파일 업로드 제외 개발 완료

+++ src/main/java/com/cn/interri/design/repository/RoomTypeRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-public interface RoomTypeRepository extends JpaRepository<RoomType, Long> {
+public interface RoomTypeRepository extends JpaRepository<RoomType, Integer> {
 
     @Query("SELECT new com.cn.interri.design.dto.RoomTypeDto(rt.id, rt.roomTypeNm) FROM RoomType rt")
     List<RoomTypeDto> findAllRoomType();
