+++ src/main/java/com/cn/interri/design/repository/StyleRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-public interface StyleRepository extends JpaRepository<Style, Long> {
+public interface StyleRepository extends JpaRepository<Style, Integer> {
 
 
     @Query("SELECT new com.cn.interri.design.dto.StyleDto(s.id, s.styleNm) FROM Style s")
